write a new test file for testing assign hearings table
